+++ mcp_server/plugins/kotlin_plugin/__init__.py
@@ -2,4 +2,6 @@
 
 from .plugin import KotlinPlugin
 
-__all__ = ["KotlinPlugin"]
+Plugin = KotlinPlugin
+
+__all__ = ["KotlinPlugin", "Plugin"]