+++ mcp_server/plugins/rust_plugin/__init__.py
@@ -2,4 +2,6 @@
 
 from .plugin import RustPlugin
 
-__all__ = ["RustPlugin"]
+Plugin = RustPlugin
+
+__all__ = ["RustPlugin", "Plugin"]